src/backend/requirements.txt
django>=5.0.0
django-cors-headers
python-dotenv
google-genai
pillow